moved tests into folders
