+++ components/application/ApplicationForm.tsx
@@ -10,7 +10,7 @@ import {
 import { Questions, Sections } from '../../common/questions';
 import { Alert, Button, Form, notification } from 'antd';
 import useSWR from 'swr';
-import { format } from '../dashboard/StatusDialogue';
+import { format } from '../dashboard/status-dialogue/StatusDialogue';
 import { isAfterRegistrationClosed, isBeforeRegistrationOpens } from '../../common/utils';
 import { useWarnIfUnsavedChanges } from '../hooks/useWarnIfUnsavedChanges';
 import { FormSectionsAndQuestions } from './FormSectionsAndQuestions';
