+++ components/dashboard/application-status-dialogue/ApplicationStatusDialogue.test.tsx
@@ -1,8 +1,8 @@
 import { describe, expect, it } from '@jest/globals';
 import { render, screen } from '@testing-library/react';
-import { ApplicationStatus } from '../common/types';
-import { RegistrationState } from '../components/hooks/useRegistrationState';
-import { ApplicationStatusDialogue } from '../components/dashboard/ApplicationStatusDialogue';
+import { ApplicationStatus } from '../../../common/types';
+import { RegistrationState } from '../../hooks/useRegistrationState';
+import { ApplicationStatusDialogue } from './ApplicationStatusDialogue';
 
 describe('Application Status Dialogue', () => {
   describe('has a submitted application status', () => {
