+++ components/dashboard/status-dialogue/StatusDialogue.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
 import { render, screen } from '@testing-library/react';
-import { RSVPStatus } from '../common/types';
-import { StatusDialogue } from '../components/dashboard/StatusDialogue';
+import { RSVPStatus } from '../../../common/types';
+import { StatusDialogue } from './StatusDialogue';
 
 describe('Rsvp Dialogue', () => {
   it('renders the ApplyLater component when registrationState is BeforeOpen', () => {
